+++ tests/AppFrameworkTestClasses/Traits/DBHelperTestTrait.php
@@ -9,8 +9,11 @@ declare(strict_types=1);
 namespace AppFrameworkTestClasses\Traits;
 
 use DBHelper;
+use TestDriver\Area\TestingScreen;
+use TestDriver\ClassFactory;
 use TestDriver\TestDBRecords\TestDBCollection;
 use TestDriver\TestDBRecords\TestDBRecord;
+use TestDriver\TestDBRecords\TestDBRecordSelectionTieIn;
 
 /**
  * @package Application
@@ -33,11 +36,18 @@ trait DBHelperTestTrait
         DBHelper::resetTrackedQueries();
     }
 
-    public function createTestRecord(?string $label=null, ?string $alias=null) : TestDBRecord
+    public function createTestDBRecord(?string $label=null, ?string $alias=null) : TestDBRecord
     {
         return TestDBCollection::getInstance()->createTestRecord(
             $label ?? 'Test Record '.$this->getTestCounter('test-records'),
             $alias ?? 'test-record-'.$this->getTestCounter('test-records')
         );
+    }
+
+    public function createTestDBRecordTieIn() : TestDBRecordSelectionTieIn
+    {
+        $screen = ClassFactory::createDriver()->getScreenByPath(TestingScreen::URL_NAME);
+
+        return new TestDBRecordSelectionTieIn($screen);
     }
 }